+++ app/src/main/java/org/breezyweather/sources/dmi/DmiService.kt
@@ -48,7 +48,7 @@ class DmiService @Inject constructor(
             .create(DmiApi::class.java)
     }
 
-    override val supportedFeaturesInMain = listOf<SecondaryWeatherSourceFeature>(
+    override val supportedFeaturesInMain = listOf(
         SecondaryWeatherSourceFeature.FEATURE_ALERT
     )