+++ CHANGELOG.md
@@ -13,7 +13,8 @@ This major version introduces new hourly trends for more detailed weather data, 
 
 **Weather sources**
 - [Danmarks Meteorologiske Institut (DMI)] Added support as a worldwide main source, with alerts for Denmark-only
-- [National Weather Service (NWS)] Added support as a main source in United States of America-only
+- [National Weather Service (NWS)] Added support as a main source in United States of America and territories-only
+- [National Weather Service (NWS)] Added support as a secondary source for alerts in United States of America and territories-only
 - [Bright Sky (DWD)] Added support as a main source in Germany-only
 - [Bright Sky (DWD)] Added support as a secondary source for alerts in Germany-only
 - [Environment and Climate Change Canada] Added support as a main source in Canada-only